+++ legends/constants.py
@@ -5,13 +5,13 @@ update to the game is released.
 
 Instructions for updating game assets: Open Star Trek: Legends and let
 it play for a while to be sure all assets download. Run
-`legends.setup.getAssets`. In Visual Studio, create an empty C# project,
+`legends.build.getAssets`. In Visual Studio, create an empty C# project,
 add reference to 'Assembly-CSharp.dll', then add the code in 'temp.cs'
-somewhere in the project. Use `legends.setup.exportData` to export the
+somewhere in the project. Use `legends.build.exportData` to export the
 current data to the current working directory. Compare the new data and
 the current data to be sure everything looks okay. Once satisfied, use
-`legends.updateData` to replace the data inside the `legends` package
-with the contents of the `data-new` folder.
+`legends.build.updateData` to replace the data inside the `legends`
+package with the contents of the `data-new` folder.
 
 Attributes:
     CHARACTER (dict): A direct parsing of `data/GSCharacter.json`
@@ -86,7 +86,7 @@ Attributes:
 """
 
 from legends.utils.functions import readData
-from legends.setup import (
+from legends.build import (
     ROOT, getItems, getSummonPool, getPartStats, getPowerFunc, getSkillIDs
 )